+++ packages/magento-cart-shipping-address/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Change Log
 
+## 8.1.0-canary.38
+
+### Minor Changes
+
+- [#2305](https://github.com/graphcommerce-org/graphcommerce/pull/2305) [`77e8297`](https://github.com/graphcommerce-org/graphcommerce/commit/77e82976816994336c616208a651cb18ce9ea270) - Add company and vat fields to shipping and billing forms
+  ([@Giovanni-Schroevers](https://github.com/Giovanni-Schroevers))
+
+### Patch Changes
+
+- [#2305](https://github.com/graphcommerce-org/graphcommerce/pull/2305) [`77e8297`](https://github.com/graphcommerce-org/graphcommerce/commit/77e82976816994336c616208a651cb18ce9ea270) - Fix bug with persist not applying saved changes by moving <FromPersist/> below the form components
+  ([@Giovanni-Schroevers](https://github.com/Giovanni-Schroevers))
+
+## 8.1.0-canary.37
+
+### Patch Changes
+
+- [#2300](https://github.com/graphcommerce-org/graphcommerce/pull/2300) [`19adebe`](https://github.com/graphcommerce-org/graphcommerce/commit/19adebe2ff896c65d21543fc9243a36393747e06) - Fixed CustomerAddressForm resetting to a saved address when trying to use a new address
+  ([@bramvanderholst](https://github.com/bramvanderholst))
+
 ## 8.1.0-canary.36
 
 ## 8.1.0-canary.35