+++ packages/magento-pagebuilder/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @graphcommerce/magento-pagebuilder
 
+## 8.1.0-canary.38
+
+## 8.1.0-canary.37
+
 ## 8.1.0-canary.36
 
 ## 8.1.0-canary.35