+++ packages/magento-payment-multisafepay/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @graphcommerce/magento-payment-multisafepay
 
+## 8.1.0-canary.38
+
+### Patch Changes
+
+- [#2305](https://github.com/graphcommerce-org/graphcommerce/pull/2305) [`77e8297`](https://github.com/graphcommerce-org/graphcommerce/commit/77e82976816994336c616208a651cb18ce9ea270) - Fix bug with persist not applying saved changes by moving <FromPersist/> below the form components
+  ([@Giovanni-Schroevers](https://github.com/Giovanni-Schroevers))
+
+## 8.1.0-canary.37
+
 ## 8.1.0-canary.36
 
 ## 8.1.0-canary.35